refactored code to make it more pleasant to my teammates
