+++ matching-service/index.js
@@ -6,10 +6,14 @@ import { newMatchHandler, disconnectHandler, leaveMatchHandler, setIo } from './
 import { MATCH_LEAVE, MATCH_REQUEST_NEW } from './public/events.js';
 
 const app = express();
+
+// Middleware
 app.use(express.urlencoded({ extended: true }))
 app.use(express.json())
 app.use(cors()) // config cors so that front-end can use
 app.options('*', cors())
+
+// App Handlers
 app.use(express.static('public'))
 const httpServer = createServer(app)
 const io = new Server(httpServer, { /* options */ });
