+++ matching-service/public/index.html
@@ -1,3 +1,4 @@
+<link rel="stylesheet" href="./styles.css">
 <div id="selection-page">
     <h1>Choose your difficulty</h1>
     <p>userid : <input id="userid" type="text" value="1"></p>
@@ -24,5 +25,4 @@
     <button id="leave">Leave</button>
 </div>
 <script src="https://cdn.socket.io/4.5.0/socket.io.min.js" integrity="sha384-7EyYLQZgWBi67fBtVxw60/OWl1kjsfrPFcaU0pp0nAh+i8FD068QogUvg85Ewy1k" crossorigin="anonymous"></script>
-<link rel="stylesheet" href="./styles.css">
 <script src='./main.js' type="module"></script>
